+++ circuitx/gesture-navigation/src/commonMain/kotlin/com/slack/circuitx/gesturenavigation/PredictiveBackNavigationDecorator.kt
@@ -14,10 +14,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.geometry.Offset
-import com.slack.circuit.backstack.NavArgument
 import com.slack.circuit.foundation.animation.AnimatedNavDecorator
 import com.slack.circuit.foundation.internal.PredictiveBackEventHandler
 import com.slack.circuit.runtime.InternalCircuitApi
+import com.slack.circuit.runtime.navigation.NavArgument
+import com.slack.circuit.runtime.navigation.NavStackList
+import com.slack.circuit.runtime.navigation.navStackListOf
 import kotlin.math.abs
 
 internal abstract class PredictiveBackNavigationDecorator<T : NavArgument>(
@@ -40,18 +42,23 @@ internal abstract class PredictiveBackNavigationDecorator<T : NavArgument>(
   protected var swipeOffset: Offset by mutableStateOf(Offset.Zero)
     private set
 
-  override fun targetState(args: List<T>): GestureNavTransitionHolder<T> {
+  override fun targetState(args: NavStackList<T>): GestureNavTransitionHolder<T> {
     return GestureNavTransitionHolder(args)
   }
 
   @OptIn(InternalCircuitApi::class)
   @Composable
-  override fun updateTransition(args: List<T>): Transition<GestureNavTransitionHolder<T>> {
+  override fun updateTransition(args: NavStackList<T>): Transition<GestureNavTransitionHolder<T>> {
     val current = remember(args) { targetState(args) }
     val previous =
       remember(args) {
-        if (args.size > 1) {
-          targetState(args.subList(1, args.size))
+        val hasBackward = args.backwardItems.iterator().hasNext()
+        if (hasBackward) {
+          // Building the state we'd go to if we go backwards.
+          val forward = listOf(args.active) + args.forwardItems
+          val current = args.backwardItems.first()
+          val backward = args.backwardItems.drop(1)
+          targetState(navStackListOf(forward, current, backward))
         } else null
       }
 
@@ -79,7 +86,7 @@ internal abstract class PredictiveBackNavigationDecorator<T : NavArgument>(
       }
     }
     PredictiveBackEventHandler(
-      isEnabled = args.size > 1,
+      isEnabled = previous != null,
       onBackProgress = { progress, offset ->
         showPrevious = progress != 0f
         swipeProgress = progress